+++ android/app/src/main/AndroidManifest.xml
@@ -21,6 +21,9 @@
         android:theme="@style/Theme.Social"
         tools:targetApi="31">
         <activity
+            android:name=".presentation.profile.ProfileActivity"
+            android:exported="false" />
+        <activity
             android:name=".OnboardingActivity"
             android:exported="false" />
         <activity
