+++ android/app/src/main/java/com/social/presentation/interaction/chats/ChatScreenFragment.kt
@@ -3,7 +3,6 @@ package com.social.presentation.interaction.chats
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
 import com.social.R
 import com.social.databinding.FragmentChatScreenBinding
 import com.social.domain.model.ChatUserData
@@ -29,7 +28,8 @@ class ChatScreenFragment : Fragment(R.layout.fragment_chat_screen) {
     private fun setupViews() {
         EmojiManager.install(GoogleEmojiProvider())
 
-        emojiPopup = EmojiPopup.Builder.fromRootView(binding.rootView).build(binding.inputSendMessage)
+        emojiPopup =
+            EmojiPopup.Builder.fromRootView(binding.rootView).build(binding.inputSendMessage)
 
         binding.btnFace.setOnClickListener {
             emojiPopup.toggle()
@@ -57,7 +57,6 @@ class ChatScreenFragment : Fragment(R.layout.fragment_chat_screen) {
         }
 
         binding.btnBack.setOnClickListener {
-            findNavController().navigate(R.id.action_chatScreenFragment_to_listChatsFragment)
         }
     }
