+++ android/app/src/main/java/com/social/presentation/profile/EditProfileFragment.kt
@@ -13,7 +13,6 @@ import android.widget.ArrayAdapter
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
 import com.social.R
 import com.social.databinding.FragmentEditProfileBinding
 import com.social.databinding.ItemSocialLinkLayoutBinding
@@ -49,7 +48,6 @@ class EditProfileFragment : Fragment(R.layout.fragment_edit_profile) {
         }
 
         binding.iconLeft.setOnClickListener {
-            findNavController().navigate(R.id.action_editProfileFragment_to_userProfileFragment)
             cleanFormEditProfile()
         }
 
@@ -120,7 +118,8 @@ class EditProfileFragment : Fragment(R.layout.fragment_edit_profile) {
     }
 
     private fun toggleLinearLayoutVisibility() {
-        val visibility = if (binding.linearLayoutContainer.visibility == View.VISIBLE) View.GONE else View.VISIBLE
+        val visibility =
+            if (binding.linearLayoutContainer.visibility == View.VISIBLE) View.GONE else View.VISIBLE
         binding.linearLayoutContainer.visibility = visibility
     }
