+++ android/app/src/main/java/com/social/presentation/profile/UserProfileFragment.kt
@@ -1,7 +1,5 @@
 package com.social.presentation.profile
 
-import android.content.Context.MODE_PRIVATE
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -11,7 +9,6 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.google.android.material.imageview.ShapeableImageView
-import com.social.OnboardingActivity
 import com.social.R
 import com.social.databinding.FragmentUserProfileBinding
 import com.social.databinding.ItemPostBinding
@@ -74,10 +71,18 @@ class UserProfileFragment : Fragment(R.layout.fragment_user_profile) {
     }
 
     private fun action() {
+        val transaction = requireActivity().supportFragmentManager.beginTransaction()
+
         binding.buttonEditProfile.setOnClickListener {
+            transaction.replace(R.id.navHostFragment, EditProfileFragment())
+            transaction.addToBackStack(null)
+            transaction.commit()
         }
+
         binding.buttonSettingProfile.setOnClickListener {
-            logout()
+            transaction.replace(R.id.navHostFragment, SettingProfileFragment())
+            transaction.addToBackStack(null)
+            transaction.commit()
         }
     }
 
@@ -148,20 +153,5 @@ class UserProfileFragment : Fragment(R.layout.fragment_user_profile) {
         imageView: ShapeableImageView,
     ) {
         Picasso.get().load(imageURL).into(imageView)
-    }
-
-    private fun logout() {
-        val preference =
-            requireContext().getSharedPreferences("login_saved", MODE_PRIVATE)
-        val editor = preference.edit()
-        editor.remove("psw")
-        editor.remove("check")
-        editor.apply()
-        startActivity(
-            Intent(
-                requireContext(),
-                OnboardingActivity::class.java,
-            ),
-        )
     }
 }
