--- android/app/src/main/res/navigation/navigation.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/navigation"
-    app:startDestination="@id/homeFragment">
-    <fragment
-        android:id="@+id/listChatsFragment"
-        android:name="com.social.presentation.interaction.chats.ListChatsFragment"
-        android:label="fragment_list_chat"
-        tools:layout="@layout/fragment_list_chat">
-        <action
-            android:id="@+id/action_listChatsFragment_to_chatScreenFragment"
-            app:destination="@id/chatScreenFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/chatScreenFragment"
-        android:name="com.social.presentation.interaction.chats.ChatScreenFragment"
-        android:label="fragment_chat_screen"
-        tools:layout="@layout/fragment_chat_screen">
-        <action
-            android:id="@+id/action_chatScreenFragment_to_listChatsFragment"
-            app:destination="@id/listChatsFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="com.social.presentation.home.HomeFragment"
-        android:label="fragment_home"
-        tools:layout="@layout/fragment_home" />
-    <fragment
-        android:id="@+id/userProfileFragment"
-        android:name="com.social.presentation.profile.UserProfileFragment"
-        android:label="fragment_user_profile"
-        tools:layout="@layout/fragment_user_profile">
-        <action
-            android:id="@+id/action_userProfileFragment_to_editProfileFragment"
-            app:destination="@id/editProfileFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/editProfileFragment"
-        android:name="com.social.presentation.profile.EditProfileFragment"
-        android:label="fragment_edit_profile"
-        tools:layout="@layout/fragment_edit_profile" >
-        <action
-            android:id="@+id/action_editProfileFragment_to_userProfileFragment"
-            app:destination="@id/userProfileFragment" />
-    </fragment>
-</navigation>
