+++ android/app/src/main/res/values/strings.xml
@@ -93,4 +93,5 @@
     <string name="social_link">Link</string>
     <string name="search">Busqueda</string>
     <string name="work">Work</string>
+    <string name="logout">Cerrar sesión</string>
 </resources>
